hypershift:performanceprofile: associate profile name with user input
When a `PerformanceProfile` should be mirrored into the HCP namespace,
its given name which compose from a constant (perfprof) + the node pool name.

In case there are multiple `PerformanceProfiles`, one per hosted cluster,
the approach mentioned above makes it hard to understand which `PreformanceProfile`
in the clusters namespace is associate with each hostedcluster.

It requires to examine the node pool \`spec.tuningConfig\` to figure out
the relation between user profile and the mirrored one.

This commit suggest to embed the configMap (that encapsulates the performanceProfile) name
in the mirrored profile, to make the relation clear and observable on
first sight.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>